Made 3rd Mini Game
Tried to add another object in the ship sailing mini-game where it goes from left to right once it collides with a collider. But I was having problems with it.
Also started making the third mini-game where the player cleans the ship. I started with spawning the spots that the player needs to clean.